Enhanced the speed of the graphics
git-svn-id: https://subversion.assembla.com/svn/sdrsharp/trunk@597 d3cde5d6-82b8-4669-8b62-43054d6781d8

+++ PanView/SpectrumAnalyzer.cs
@@ -362,7 +362,7 @@ namespace SDRSharp.PanView
                 {
                     var carrierPen = redPen;
                     carrierPen.Width = CarrierPenWidth;
-                    graphics.FillRectangle(transparentBackground, _lower, 0, bandpassWidth, ClientRectangle.Height);
+                    graphics.FillRectangle(transparentBackground, (int) _lower, 0, (int) bandpassWidth, ClientRectangle.Height);
                     if (xCarrier >= AxisMargin && xCarrier <= ClientRectangle.Width - AxisMargin)
                     {
                         graphics.DrawLine(carrierPen, xCarrier, 0f, xCarrier, ClientRectangle.Height);
